More styling changes, delted sidebar styling file and moved contents to main styles.scss
